convert float types to double

+++ Julia_Morph/julia.cuh
@@ -22,13 +22,13 @@ void julia(void);
 void initColors(sf::Uint8* h_colorTable, int max_iter);
 
 __global__ void cudaJulia(int w, int h, sf::Uint8* d_counts, sf::Uint8* d_colorTable,
-    int const max_iter, float re_min, float im_min, float re_scale, float im_scale,
-    float P, float Q, int setType, float z, float x, float c);
+    int const max_iter, double re_min, double im_min, double re_scale, double im_scale,
+    double P, double Q, int setType, double z, double x, double c);
 
-__device__ int mandelbrot(int iter, int max_iter, float A, float B, float P, float Q);
+__device__ int mandelbrot(int iter, int max_iter, double A, double B, double P, double Q);
 
-__device__ int burningShip(int iter, int max_iter, float A, float B, float P, float Q);
+__device__ int burningShip(int iter, int max_iter, double A, double B, double P, double Q);
 
-__device__ int mandelCubed(int iter, int max_iter, float A, float B, float P, float Q);
+__device__ int mandelCubed(int iter, int max_iter, double A, double B, double P, double Q);
 
-__device__ int experimental(int iter, int max_iter, float A, float B, float s, float r, float f);
+__device__ int experimental(int iter, int max_iter, double A, double B, double s, double r, double f);
